+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## [Unreleased]
 * Move documentation to packages readme.
+* Setup trusted publishing.
 
 ## [0.4.0] - 2025-12-06
 * minor: use live config values to better support VSCode extension.