+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 > This project follows the [Keep a Changelog](https://keepachangelog.com/en/1.1.0/) format.
 
 ## [Unreleased]
+
+## [0.4.1] - 2026-03-09
 * Move documentation to packages readme.
 * Setup trusted publishing.
 
@@ -58,5 +60,7 @@
 [0.2.0]: https://github.com/vitalets/global-cache/compare/0.1.2...0.2.0
 [0.1.2]: https://github.com/vitalets/global-cache/compare/0.1.1...0.1.2
 
-[unreleased]: https://github.com/vitalets/global-cache/compare/0.4.0...HEAD
+[Unreleased]: https://github.com/vitalets/global-cache/compare/0.4.1...HEAD
 [0.4.0]: https://github.com/vitalets/global-cache/compare/0.4.0-4...0.4.0
+
+[0.4.1]: https://github.com/vitalets/global-cache/releases/tag/0.4.1